Cast (int) ITEM.get(ChoiceGroup)

+++ src/OpenTTY.java
@@ -4124,8 +4124,8 @@ class Lua {
                                 else if (item instanceof Gauge) { formData.addElement(new Double(((Gauge) item).getValue())); }
                                 else if (item instanceof ChoiceGroup) {
                                     ChoiceGroup cg = (ChoiceGroup) item;
-                                    
-                                    if (ITEM.get(cg) == Choice.MULTIPLE) {
+
+                                    if ((int) ITEM.get(cg) == Choice.MULTIPLE) {
                                         Hashtable selTable = new Hashtable();
                                         for (int j = 0; j < cg.size(); j++) { selTable.put(new Double(j + 1), new Boolean(cg.isSelected(j))); }
